Move switch actions to separate file
This is needed so we don't break the APC (configurator)
when we extend the switch actions.

Also the action definitions just don't belong in config.h
